Don’t assume ownership of Geant4-owned raw pointers in utr-owned shared_ptrs
Since 08ef9099764f9c77a30eccbf7339ec6269f3ed64, shared_ptr is used in
all sensitive detectors to encapsulate raw pointers returned by Geant4.

The usage of smart pointers is a noble endveavour worthy of support,
however, the internals of Geant4 make it difficult to persue this
paradigm, because it relies on its own internal memory management and
only exposes raw pointers without any ownership information externally.
Trying to shoehorn these raw pointers into shared pointers is a recipe for
desaster, because as soon as the owning object goes out of scope, the
underlying raw pointer is deleted, with Geant4 being unaware of this
change. This inevitably results in attempting to access a null pointer
or potentially even double-freeing it.

The hit vector in the TupleManager is especially weird, because it is
used both by objects created by us but also by Geant4. Since we cannot
change Geant4, we have to build or own interfaces and similarly pass raw
pointers. But at least we can manage the ownership of the objects we
create using smart pointers, only requireing an additional copy to
(safely) pass the raw pointer along to the internal interface.